+++ pkg/constant/constant.go
@@ -11,8 +11,8 @@ const (
 	// Transaction sent to the TSS or ERC20 Custody address containing this message are considered as a donation
 	DonationMessage = "I am rich!"
 
-	// CmdWhitelistERC20 is used for CCTX of type cmd to give the instruction to the TSS to whitelist an ERC20 on an exeternal chain
-	CmdWhitelistERC20 = "cmd_whitelist_erc20"
+	// CmdWhitelistAsset is used for CCTX of type cmd to give the instruction to the TSS to whitelist an ERC20 on an exeternal chain
+	CmdWhitelistAsset = "cmd_whitelist_asset"
 
 	// CmdMigrateTSSFunds is used for CCTX of type cmd to give the instruction to the TSS to transfer its funds on a new address
 	CmdMigrateTSSFunds = "cmd_migrate_tss_funds"