+++ x/authority/keeper/migrator.go
@@ -5,6 +5,7 @@ import (
 
 	v4 "github.com/zeta-chain/node/x/authority/migrations/v4"
 	v5 "github.com/zeta-chain/node/x/authority/migrations/v5"
+	v6 "github.com/zeta-chain/node/x/authority/migrations/v6"
 )
 
 // Migrator is a struct for handling in-place store migrations.
@@ -27,4 +28,9 @@ func (m Migrator) Migrate3to4(ctx sdk.Context) error {
 // Migrate4to5 migrates the authority store from consensus version 4 to 5
 func (m Migrator) Migrate4to5(ctx sdk.Context) error {
 	return v5.MigrateStore(ctx, m.authorityKeeper)
+}
+
+// Migrate5to6 migrates the authority store from consensus version 5 to 6
+func (m Migrator) Migrate5to6(ctx sdk.Context) error {
+	return v6.MigrateStore(ctx, m.authorityKeeper)
 }