+++ x/authority/types/authorization_list_test.go
@@ -420,7 +420,7 @@ func TestDefaultAuthorizationsList(t *testing.T) {
 		var AdminPolicyMessageList = []string{
 			sdk.MsgTypeURL(&crosschaintypes.MsgMigrateTssFunds{}),
 			sdk.MsgTypeURL(&crosschaintypes.MsgUpdateTssAddress{}),
-			sdk.MsgTypeURL(&crosschaintypes.MsgWhitelistERC20{}),
+			sdk.MsgTypeURL(&crosschaintypes.MsgWhitelistAsset{}),
 			sdk.MsgTypeURL(&fungibletypes.MsgDeployFungibleCoinZRC20{}),
 			sdk.MsgTypeURL(&fungibletypes.MsgUpdateContractBytecode{}),
 			sdk.MsgTypeURL(&fungibletypes.MsgUpdateSystemContract{}),