+++ changelog.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Breaking Changes
+
+* The message `MsgWhitelistERC20` has been renamed to `MsgWhitelistAsset`. The message parameters remain unchanged.
+  * The event `EventERC20Whitelist` has been renamed to `EventAssetWhitelist`. The event parameters remain unchanged.
+
 ### Features
 
 * [4064](https://github.com/zeta-chain/node/pull/4064) - add support for withdraws using the new v2 connector contract
@@ -14,6 +19,7 @@
 * [4070](https://github.com/zeta-chain/node/pull/4070) - remove support for v1 revert address for BTC
 * [4144](https://github.com/zeta-chain/node/pull/4144) - standardize structured logging for zetaclient
 * [4180](https://github.com/zeta-chain/node/pull/4180) - remove unused loggers and log fields
+* [4203](https://github.com/zeta-chain/node/pull/4203) - rename `whitelistERC20` into `whitelistAsset`
 * [4199](https://github.com/zeta-chain/node/pull/4199) - remove `MsgUpdateERC20CustodyPauseStatus` and `MsgMigrateERC20CustodyFunds`
 * [4205](https://github.com/zeta-chain/node/pull/4205) - remove index field in ballot
 * [4200](https://github.com/zeta-chain/node/pull/4200) - remove `LastBlockHeight` state variable